moved with_assertions to chain_generics
