+++ zingo-testutils/src/chain_generics.rs
@@ -13,3 +13,5 @@
 pub mod conduct_chain;
 
 pub mod fixtures;
+
+pub mod with_assertions;
