+++ zingo-testutils/src/lightclient.rs
@@ -117,5 +117,3 @@ pub mod from_inputs {
         proposer.propose_send(request).await
     }
 }
-
-pub mod with_assertions;
